feat: frontend filters params, fix tests
