+++ frontend/src/pages/Tasks/ui/Tasks.tsx
@@ -1,3 +1,4 @@
+import { useMemo } from 'react';
 import { useSearchParams } from 'react-router';
 import { useNotepad } from '@pages/Tasks/lib';
 import { ErrorFetching } from '@shared/ui';
@@ -5,16 +6,28 @@ import { TasksHeader } from './TasksHeader';
 import { TasksBody } from './TasksBody';
 import { TasksSkeleton } from './TasksSkeleton';
 import { taskQueryParamsSchema } from 'shared/schemas';
+import { extractInvalidKeys } from '@sharedCommon/';
 
 export const Tasks = () => {
   const { title, notepadId, location, isError, isLoading } = useNotepad();
   const [searchParams, setSearchParams] = useSearchParams();
 
-  const validation = taskQueryParamsSchema.safeParse(
-    Object.fromEntries(searchParams),
-  );
+  const { validParams } = useMemo(() => {
+    const validation = taskQueryParamsSchema.safeParse(
+      Object.fromEntries(searchParams),
+    );
 
-  const params = validation.success ? searchParams : new URLSearchParams();
+    if (validation.success) {
+      return { validParams: searchParams };
+    }
+
+    const paramsCopy = new URLSearchParams(searchParams);
+    const invalidKeys = extractInvalidKeys(validation.error);
+
+    invalidKeys.forEach(key => paramsCopy.delete(key));
+
+    return { validParams: paramsCopy };
+  }, [searchParams]);
 
   if (isLoading) {
     return <TasksSkeleton />;
@@ -30,7 +43,7 @@ export const Tasks = () => {
       <TasksBody
         notepadId={notepadId}
         notepadPathName={location}
-        params={params}
+        params={validParams}
       />
     </>
   );
