+++ shared/index.ts
@@ -1,2 +1,3 @@
 export * from './routes';
 export * from './schemas';
+export * from './utils';
